@@ -6,36 +6,38 @@ authors = ["Marc Rijken <marc@rijken.org>"]
 license = "MIT"
 repository = "https://github.com/mrijken/exposure"
 readme = "README.md"
-packages = [
-    { include = "exposure" },
-]
+packages = [{ include = "exposure" }]
 [tool.poetry.dependencies]
-python = ">=3.6,<4.0"
+python = ">=3.10,<4.0"
 
-[tool.poetry.dev-dependencies]
-black = "^20.8b1"
-pylint = "^2.6.0"
+[tool.poetry.group.dev.dependencies]
 pytest-cov = ">=2.10.1"
 mypy = ">=0.790"
-isort = "^5.6.4"
-
-[tool.black]
-line_length = 120
-
-[tool.pylint.format]
-max-line-length = "120"
-
-[tool.isort]
-profile = "black"
-line_length = 120
-known_first_party = ["exposure"]
+ruff = "^0.1.9"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
 addopts = "--doctest-modules --doctest-glob=\"*.md\""
-testpaths = [
-    "."
-]
+testpaths = ["."]
+
+[tool.mypy]
+explicit_package_bases = true
+namespace_packages = true
+ignore_missing_imports = true
+
+[tool.ruff]
+target-version = "py310"
+line-length = 120
+select = ["ALL"]
+ignore = ["D", "ANN102", "ANN101", "EM102", "TCH", "TRY003", "EM", "FBT"]
+
+[tool.ruff.per-file-ignores]
+"tests/**" = ["S101", "INP", "SLF", "ARG", "PLR2004", "ANN201"]
+"notebooks/**" = ["ALL"]
+
+[tool.ruff.isort]
+known-first-party = ["dvb"]
+force-single-line = true
 
 [tool.vulture]
 exclude = [".eggs", ".git", ".*cache", ".tox"]
